docs: add more explanations to config options
I noticed that explanations to some config options did
not tell the user much. E.g. "Whether to enable MAAS HA"
does not give additional info for the `enable_maas_ha` option.
The PR tries to make it easier for people to understand what
the options change.

Can be even more improved for the `maas-config` part.
This could explain potential options or link to docs.